+++ packages/ccdi-server/src/responses/by/count.rs
@@ -1,4 +1,5 @@
 //! Responses for grouping by fields and counting them.
 
+pub mod file;
 pub mod sample;
 pub mod subject;